+++ src/admin/components/utilities/Auth/types.ts
@@ -1,4 +1,4 @@
-import { User } from 'config/types';
+import { User } from '@shared/types';
 
 export type AuthContext<T = User> = {
   user?: T | null;
