+++ src/admin/components/utilities/DocumentInfo/types.tsx
@@ -1,4 +1,4 @@
-import { Field } from 'config/types';
+import { Field } from '@shared/types';
 
 export type ContextType = {
   label: string;
