+++ src/shared/types.ts
@@ -1,0 +1,13 @@
+import { Type } from '@admin/components/elements/Table/Cell/types';
+
+export type Field = {
+  field: string;
+  label: string;
+  type: typeof Type[keyof typeof Type];
+};
+
+export type User = {
+  id: number;
+  email: string;
+  userName: string;
+};
